pyproject.toml
[project]
name = "fastapi-async-way-pg-with-todo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.118.0",
    "greenlet>=3.2.4",
    "passlib>=1.7.4",
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.11.0",
    "python-json-logger>=4.0.0",
    "sqlmodel>=0.0.27",
]



[dependency-groups]
dev = ["ruff>=0.13.3"]

[tool.ruff]
src = ["app"]
line-length = 88
target-version = "py312"
exclude = [ "migrations",      # Alembic migration scripts
    "tests",           # Test files (optional)
    ".venv",           # Virtual environment
    "build",           # Build artifacts
    "dist",            # Distribution packages
    "__pycache__",     # Cache folders]
]

[tool.ruff.lint]
select=["E","F","W","I"]
ignore=["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
combine-as-imports = true
force-sort-within-sections = true
lines-after-imports = 2
order-by-type = true

[tool.ruff.format]
quote-style="single"
indent-style="space"


